Added new Packer::ElementType:E_INT128,E_DOUBLe80,E_RATIONAL,E_ARRAY. removed E_INT (called E_RESERVED). E_INT is the3 same as E_LONG All operators<</>> for Packer makes a check and reduce size of variable if possible before packing. Modified encoding of BigReal into Packer, using BCDArray, which automaticly takes careof any endianess Added isFloat,isDouble,isDouble80, getFloatNoLimieCheck, getDoubleNoLimitCheck to BigReal Added isInt,isUint,isInt64,isUint64,isInt128,isUint128 for arg BigReal/Double80 Moved most functions in Math/Double64.h to Util/MathUtil.h, Moved many math-functions declared in MyUtil.h to MathUtil.h. this file is included from MyUtil.h Added operator short, unsigned short, char, unsigned char to _int128/_uint128 Added opetaor ==,!=,<,>,<=,>= for (_int128/_uint128,char/unsigned char) Added isChar,-Uchar,-Short,-Ushort,-Int,-Uint,-Int64,-Uint64 for arg _int128,/_uint128 Added static const Rational Rational::zero, -::one, operator++, and --, prefix and postfix form. Optimized constructor Rational(float/double/Double80) Moved all versions of sprintbin to MyString.h Added macros __ORDER_BIG_ENDIAN__/__ORDER_LISTLE_ENDIAN__ and defined __BYTE_ORDER__ to be defined as __ORDER_LITTLE_ENDIAN__ Changed all checkes #if __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__ ->                     #if __BYTE_ORDER__ != __ORDER_BIG_ENDIAN__ Modified XMLDoc::setValue for floating points, using int/int64 whenever possible Added unittests for Packer  - test of Packer in TestColleciton (for Map/Set)  - test of Packer in TestArray  - test of Packer in BitSet

+++ Lib/Include/CompactArray.h
@@ -589,9 +589,9 @@ public:
 };
 
 template<class S, class T, class D=StreamDelimiter> S &operator<<(S &out, const CompactArray<T> &a) {
-  const D delimiter;
-  const UINT elemSize = sizeof(T);
-  const UINT64 n = a.size();
+  const D      delimiter;
+  const UINT   elemSize = sizeof(T);
+  const size_t n        = a.size();
   out << elemSize << delimiter << n << delimiter;
   if(n) {
     const T *e = a.getBuffer();
